+++ src/components/Footer.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Footer() {
+  return (
+    <div>Footer</div>
+  )
+}
+
+export default Footer
