+++ src/utils/firebaseConfig.js
@@ -1,0 +1,10 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyB1lhXzkwzslSudFPed58CVXr9i6JRLNqE",
+    authDomain: "pouch-90c28.firebaseapp.com",
+    projectId: "pouch-90c28",
+    storageBucket: "pouch-90c28.appspot.com",
+    messagingSenderId: "1055575224621",
+    appId: "1:1055575224621:web:f7bdc9b4e8b5c7b46b63a4",
+  };
+  
+  export default firebaseConfig;
